Extract esp-idf processing into its own module, create simple CLI interface
