+++ .gitignore
@@ -1,3 +1,3 @@
 /target
 **/*.rs.bk
-esp32.svd
+*.svd
