+++ Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-regex = "1.3.1"
+clap = { git = "https://github.com/clap-rs/clap/" }
+regex = "1.3.6"
 svd-parser = { git = "https://github.com/rust-embedded/svd.git", branch = "build", features = ["unproven"] }
-xmltree = "0.8.0"
+xmltree = "0.10.0"
