+++ src/idf/mod.rs
@@ -1,21 +1,28 @@
-use regex::Regex;
 use std::collections::HashMap;
 use std::fs::File;
 use std::io::prelude::*;
+use std::io::BufWriter;
 use std::ops::RangeInclusive;
 use std::str::FromStr;
 
-/* Regex's to find all the peripheral addresses */
-pub const REG_BASE: &'static str = r"\#define[\s*]+DR_REG_(.*)_BASE[\s*]+0x([0-9a-fA-F]+)";
-pub const REG_DEF: &'static str = r"\#define[\s*]+([^\s*]+)_REG[\s*]+\(DR_REG_(.*)_BASE \+ (.*)\)";
-pub const REG_DEF_INDEX: &'static str =
+use regex::Regex;
+use svd_parser::{
+    addressblock::AddressBlock, bitrange::BitRangeType, cpu::CpuBuilder, device::DeviceBuilder,
+    encode::Encode, endian::Endian, fieldinfo::FieldInfoBuilder, peripheral::PeripheralBuilder,
+    registerinfo::RegisterInfoBuilder, Access, BitRange, Device as SvdDevice, Field,
+    Register as SvdRegister, RegisterCluster,
+};
+
+const SOC_BASE_PATH: &'static str = "esp-idf/components/soc/esp32/include/soc/";
+
+// Regexes to find all the peripheral addresses
+const REG_BASE: &'static str = r"\#define[\s*]+DR_REG_(.*)_BASE[\s*]+0x([0-9a-fA-F]+)";
+const REG_DEF: &'static str = r"\#define[\s*]+([^\s*]+)_REG[\s*]+\(DR_REG_(.*)_BASE \+ (.*)\)";
+const REG_DEF_INDEX: &'static str =
     r"\#define[\s*]+([^\s*]+)_REG\(i\)[\s*]+\(REG_([0-9A-Za-z_]+)_BASE[\s*]*\(i\) \+ (.*?)\)";
-pub const REG_BITS: &'static str =
-    r"\#define[\s*]+([^\s*]+)_(S|V)[\s*]+\(?(0x[0-9a-fA-F]+|[0-9]+)\)?";
-pub const REG_BIT_INFO: &'static str =
-    r"/\*[\s]+([0-9A-Za-z_]+)[\s]+:[\s]+([0-9A-Za-z_/]+)[\s]+;bitpos:\[(.*)\][\s];default:[\s]+(.*)[\s];[\s]\*/";
-pub const REG_DESC: &'static str = r"\*description:\s(.*[\n|\r|\r\n]?.*)\*/";
-pub const INTERRUPTS: &'static str =
+const REG_BIT_INFO: &'static str = r"/\*[\s]+([0-9A-Za-z_]+)[\s]+:[\s]+([0-9A-Za-z_/]+)[\s]+;bitpos:\[(.*)\][\s];default:[\s]+(.*)[\s];[\s]\*/";
+const REG_DESC: &'static str = r"\*description:\s(.*[\n|\r|\r\n]?.*)\*/";
+const INTERRUPTS: &'static str =
     r"\#define[\s]ETS_([0-9A-Za-z_/]+)_SOURCE[\s]+([0-9]+)/\*\*<\s([0-9A-Za-z_/\s,]+)\*/";
 
 #[derive(Debug, Default, Clone)]
@@ -24,6 +31,7 @@ pub struct Peripheral {
     pub address: u32,
     pub registers: Vec<Register>,
 }
+
 #[derive(Clone, Debug, Default)]
 pub struct Interrupt {
     pub name: String,
@@ -74,8 +82,6 @@ impl Default for Bits {
     }
 }
 
-use svd_parser::Access;
-
 #[derive(Debug, Copy, Clone)]
 pub enum Type {
     // ReadAsZero,
@@ -125,7 +131,7 @@ enum State {
     CheckEnd(String, Register),
 }
 
-pub fn parse_idf(path: &str) -> HashMap<String, Peripheral> {
+fn parse_idf() -> HashMap<String, Peripheral> {
     let mut peripherals = HashMap::new();
     let mut invalid_peripherals = vec![];
     let mut invalid_files = vec![];
@@ -134,7 +140,7 @@ pub fn parse_idf(path: &str) -> HashMap<String, Peripheral> {
 
     let mut interrupts = vec![];
 
-    let filname = path.to_owned() + "soc.h";
+    let filname = SOC_BASE_PATH.to_owned() + "soc.h";
     let re_base = Regex::new(REG_BASE).unwrap();
     let re_reg = Regex::new(REG_DEF).unwrap();
     let re_reg_index = Regex::new(REG_DEF_INDEX).unwrap();
@@ -179,7 +185,7 @@ pub fn parse_idf(path: &str) -> HashMap<String, Peripheral> {
         peripherals.insert(peripheral.to_string(), p);
     }
 
-    std::fs::read_dir(path)
+    std::fs::read_dir(SOC_BASE_PATH)
         .unwrap()
         .filter_map(Result::ok)
         .filter(|f| f.path().to_str().unwrap().ends_with("_reg.h"))
@@ -347,4 +353,111 @@ fn file_to_string(fil: &str) -> String {
     let mut data = String::new();
     soc.read_to_string(&mut data).unwrap();
     data
+}
+
+fn build_svd(peripherals: HashMap<String, Peripheral>) -> Result<SvdDevice, ()> {
+    let mut svd_peripherals = vec![];
+
+    for (name, p) in peripherals {
+        let mut registers = vec![];
+        for r in p.registers {
+            let mut fields = vec![];
+            for field in &r.bit_fields {
+                let description = if field.description.trim().is_empty() {
+                    None
+                } else {
+                    Some(field.description.clone())
+                };
+
+                let bit_range = match &field.bits {
+                    Bits::Single(bit) => BitRange {
+                        offset: u32::from(*bit),
+                        width: 1,
+                        range_type: BitRangeType::OffsetWidth,
+                    },
+                    Bits::Range(r) => BitRange {
+                        offset: u32::from(*r.start()),
+                        width: u32::from(r.end() - r.start() + 1),
+                        range_type: BitRangeType::OffsetWidth,
+                    },
+                };
+
+                let field_out = FieldInfoBuilder::default()
+                    .name(field.name.clone())
+                    .description(description)
+                    .bit_range(bit_range)
+                    .build()
+                    .unwrap();
+                fields.push(Field::Single(field_out));
+            }
+
+            let info = RegisterInfoBuilder::default()
+                .name(r.name.clone())
+                .description(Some(r.description.clone()))
+                .address_offset(r.address)
+                .size(Some(32))
+                .reset_value(Some(r.reset_value as u32))
+                .fields(Some(fields))
+                .build()
+                .unwrap();
+
+            registers.push(RegisterCluster::Register(SvdRegister::Single(info)));
+        }
+        let block_size = registers.iter().fold(0, |sum, reg| {
+            sum + match reg {
+                RegisterCluster::Register(r) => r.size.unwrap(),
+                _ => unimplemented!(),
+            }
+        });
+        let out = PeripheralBuilder::default()
+            .name(name.to_owned())
+            .base_address(p.address)
+            .registers(Some(registers))
+            .address_block(Some(AddressBlock {
+                offset: 0x0,
+                size: block_size, // TODO what about derived peripherals?
+                usage: "registers".to_string(),
+            }))
+            .build()
+            .unwrap();
+
+        svd_peripherals.push(out);
+    }
+    println!("Len {}", svd_peripherals.len());
+
+    let cpu = CpuBuilder::default()
+        .name("Xtensa LX6".to_string())
+        .revision("1".to_string())
+        .endian(Endian::Little)
+        .mpu_present(false)
+        .fpu_present(true)
+        // according to https://docs.espressif.com/projects/esp-idf/en/latest/api-reference/system/intr_alloc.html#macros
+        // 7 levels so 3 bits? //TODO verify
+        .nvic_priority_bits(3)
+        .has_vendor_systick(false)
+        .build()
+        .unwrap();
+
+    let device = DeviceBuilder::default()
+        .name("Espressif".to_string())
+        .version(Some("1.0".to_string()))
+        .schema_version(Some("1.0".to_string()))
+        // broken see: https://github.com/rust-embedded/svd/pull/104
+        // .description(Some("ESP32".to_string()))
+        // .address_unit_bits(Some(8))
+        .width(Some(32))
+        .cpu(Some(cpu))
+        .peripherals(svd_peripherals)
+        .build()
+        .unwrap();
+
+    Ok(device)
+}
+
+pub fn create_svd() {
+    let peripherals = parse_idf();
+    let svd = build_svd(peripherals).unwrap();
+
+    let f = BufWriter::new(File::create("esp32.svd").unwrap());
+    svd.encode().unwrap().write(f).unwrap();
 }
